pyproject.toml
[project]
name = "ai-roaster"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "azure-ai-inference>=1.0.0b9",
    "google-generativeai>=0.8.5",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.0",
    "streamlit>=1.45.1",
]
